% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcert.R
\name{get_cert}
\alias{get_cert}
\title{Retrieve a certificate}
\usage{
get_cert(id, sha256, host = get_obs_host(), timeout = 60)
}
\arguments{
\item{id}{the ID of the certificate to start the path at}

\item{sha256}{the hexadecimal checksum of the DER certificate (only if \code{id} is not provided)}

\item{host}{if you host your own instance of the \href{https://github.com/mozilla/tls-observatory}{TLS Observatory}, either specify
the \emph{hostname} here (not advised) or set the envrionment variable \code{TLS_OBSERVATORY_HOST}
to the hostname (preferred). If no environment variable is set the default
("\code{tls-observatory.services.mozilla.com}") will be used.}

\item{timeout}{defaults to \code{60}, but most calls to this API enpoint come back quickly.
Adjust as necessary (ie. when the server is under load this might need to be increased).}
}
\value{
a \code{list} containing the parsed certificate and its raw X509 version encoded with base64
alog with an X509 certificate object created using \code{openssl}.
}
\description{
Retrieve a certificate
}
\note{
use \emph{either} \code{id} or \code{sha256}. If both are provided, \code{id} gets preference.
}
\examples{
str(get_cert(1))
}
\references{
\url{https://github.com/mozilla/tls-observatory#get-apiv1certificate}
}
