% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartbeat.R
\name{get_heartbeat}
\alias{get_heartbeat}
\title{Check for TLS Obsercatory API server online status}
\usage{
get_heartbeat(host = get_obs_host())
}
\arguments{
\item{host}{if you host your own instance of the \href{https://github.com/mozilla/tls-observatory}{TLS Observatory}, either specify
the \emph{hostname} here (not advised) or set the envrionment variable \code{TLS_OBSERVATORY_HOST}
to the hostname (preferred). If no environment variable is set the default
("\code{tls-observatory.services.mozilla.com}") will be used.}
}
\value{
\code{TRUE} if the server is up
}
\description{
Check for TLS Obsercatory API server online status
}
\examples{
get_heartbeat()
}
\references{
\url{https://github.com/mozilla/tls-observatory#get-apiv1heartbeat}
}
