% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{get_scan_results}
\alias{get_scan_results}
\title{Retrieve a certificate}
\usage{
get_scan_results(id, sha256, host = get_obs_host(), timeout = 60)
}
\arguments{
\item{id}{the ID of the certificate to start the path at}

\item{host}{if you host your own instance of the \href{https://github.com/mozilla/tls-observatory}{TLS Observatory}, either specify
the \emph{hostname} here (not advised) or set the envrionment variable \code{TLS_OBSERVATORY_HOST}
to the hostname (preferred). If no environment variable is set the default
("\code{tls-observatory.services.mozilla.com}") will be used.}

\item{timeout}{defaults to \code{60}, but most calls to this API enpoint come back quickly.
Adjust as necessary (ie. when the server is under load this might need to be increased).}
}
\value{
a \code{list} containing the scan results and the ID of the end-entity certificate.
}
\description{
Retrieve a certificate
}
\examples{
str(get_scan_results("12302333"))
}
\references{
\url{https://github.com/mozilla/tls-observatory#get-apiv1results}
}
