% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stasts.R
\name{get_usage_stats}
\alias{get_usage_stats}
\title{Retrive 24 hour scan statistics from the  TLS Obsercatory API}
\usage{
get_usage_stats(host = get_obs_host(), timeout = 60)
}
\arguments{
\item{host}{if you host your own instance of the \href{https://github.com/mozilla/tls-observatory}{TLS Observatory}, either specify
the \emph{hostname} here (not advised) or set the envrionment variable \code{TLS_OBSERVATORY_HOST}
to the hostname (preferred). If no environment variable is set the default
("\code{tls-observatory.services.mozilla.com}") will be used.}

\item{timeout}{defaults to \code{60} seconds as the endpoint seems to issue a SQL query
for every call vs cache data. Adjust as necessary (ie. when the server is under load
this might need to be increased).}
}
\value{
a \code{list} with a \code{data.frame} of scan counts for the past 24 hours and
other metadata fields.
}
\description{
Retrive 24 hour scan statistics from the  TLS Obsercatory API
}
\examples{
get_usage_stats()
}
\references{
\url{https://github.com/mozilla/tls-observatory#get-apiv1stats}
}
