% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlsobs-package.R
\docType{package}
\name{tlsobs}
\alias{tlsobs}
\alias{tlsobs-package}
\title{Tools to Work with the 'Mozilla' 'TLS' Observatory 'API'}
\description{
The Mozilla [TLS Observatory](https://observatory.mozilla.org/analyze.html) can be
used to compare your site against the Mozilla guidelines. You can use their service
or [host your own](https://github.com/mozilla/tls-observatory#getting-started). Both
the central service and locally hosted services have an API. Either the central service
 or a local service can be queried from the functions provided.
}
\author{
Bob Rudis (bob@rud.is)
}
